Implemented composite task and fixed periodic task subtle issues with periodic task running and waiting. Fixed by essentially lying to the underlying call which tick is being called.
composite task implemented, but issues with what is defined as "aborted" came up.
Ultimately decided that if all internal tasks are aborted, the composite task is aborted.
Shouldn't come up in project submission, but that's something that would have to be fixed
in an actual project.

Hopefully this will get a 100 on part 2.