Implemented composite task and fixed periodic task Subtle issue with periodic task meant it was trying to run the base task outside of its normal time frame. Fine for simple, single interval base tasks, but when combined with deadlines and other decorators things got messy. Solved by essentially lying to the base task which tick the simulation is at.
Composite task implemented as well. Minor issue with the definition of "aborted" for a composite.
As is, if all underlying tasks are aborted, the composite is aborted. No direction given for this
so it's left as is. If there is an issue in gradescope regarding this, I will adjust as necessary

Hopefully this will get me a 100 though. Fingers crossed.